% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSetup.R
\name{ModelSetup}
\alias{ModelSetup}
\title{ModelSetup: Setup the index matrix of variables of candidate models}
\usage{
ModelSetup(y, X, modeltype = "nested", intercept = FALSE)
}
\arguments{
\item{y}{a vector consist of natural numbers.}

\item{X}{a numeric matrix.}

\item{modeltype}{a string chose from 'nested' or 'all'. 'nested' means
generating candidate models in nested from according to cor(y,x); 'all' means considering
all possible combinations.}

\item{intercept}{a bool. It means the first column of X is 1, or not.}
}
\value{
an index matrix of variables of candidate models. The sth row contains
the indexes of variables used in the sth candidate model.
}
\description{
ModelSetup is a function to generate an index matrix of variables of candidate models.
The sth row contains the indexes of variables used in the sth candidate model.
}
\examples{
library(mvtnorm)
library(magrittr)
beta0=rep(0.5,5)
X=rmvnorm(100,mean=rep(0,5))
lambda=exp(X\%*\%beta0)
y=rpois(100,lambda)
ModelSetup(y,X,'nested',intercept=FALSE)


beta0=rep(0.5,5)
X=cbind(1,rmvnorm(100,mean=rep(0,4)))
lambda=exp(X\%*\%beta0)
y=rpois(100,lambda)
ModelSetup(y,X,'nested',intercept=TRUE)



}
