% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonMA.R
\name{PoissonMA}
\alias{PoissonMA}
\title{PoissonMA: Model averaging method for Poisson regressions}
\usage{
PoissonMA(y, X, Index = NULL, modeltype, intercept = intercept)
}
\arguments{
\item{y}{a vector consist of natural numbers.}

\item{X}{a numeric matrix.}

\item{Index}{NULL or an index matrix generated by ModelSetup(). Index will be calculate by ModelSetup() when Index=NULL.
If Index is assigned, modeltype and intercept are needless.}

\item{modeltype}{a string chose from 'nested' or 'all'. 'nested' means
generating candidate models in nested from according to cor(y,x); 'all' means considering
all possible combinations.}

\item{intercept}{a bool. It means the first column of X is 1, or not.}
}
\value{
a list contains model-averaging weight, optimal model index of AIC, BIC; MSE of MA, AIC, BIC, SAIC, SBIC, Full;
KL loss of MA, AIC, BIC, SAIC, SBIC, Full.
}
\description{
PoissonMA() is a function to realize model averaging method for Poisson regressions.
}
\examples{
library(mvtnorm)
beta0=rep(0.5,5)
X=rmvnorm(100,mean=rep(0,5))
lambda=exp(X\%*\%beta0)
y=rpois(100,lambda)
PoissonMA(y,X, modeltype='nested',intercept=F)


X=cbind(1,rmvnorm(100,mean=rep(0,4)))
lambda=exp(X\%*\%beta0)
y=rpois(100,lambda)
PoissonMA(y,X, modeltype='nested',intercept=T)



}
